Did the basic stuff for each class
Created the necessary java files. One for each of the classes that are said to implement. Copied the given code template to the files so that we have a guide. From the "User" class we changed the array list which was assigned to the attribute "reviews" to accept "Review" type objects instead of Strings. Added getters to all the classes that have protected/private attributes. Overriden the toString method to each class so it returns a string with all the attributes of the class. Changed the implementation of the printDetails method by using System.out.println which uses the now overriden toString method. Added comments to each part of the class.