DAY 3
Added a new attribute `allMovies` to the class `Movie` which stores all the movies in an array. Made a getter for it and also made the constructor add each new movie object created to `allMovies`. Changed the `printDetails()` methods to work like they used to, without the `toString()`. Changed the `toString()` method for each class to display less things so its easier to understand what is displayed to the terminal. Created the static method `getHighestRatedByGenre()` in the class `Movie` which returns a hashmap with genres(String) as keys and the highest rated movie of the correspoding genre as value. Used the getters of the attributes instead of the attributes by themselves that we have used in some methods. Changed the method `getRelatedMovies()` so now it fills the array `relatedMovies` with movies of the same genre and returns it. Created the method `getReviewers()` which returns an array of users who have reviewed the movie that called the method.