DAY 4
Overloaded the method `getHighestRatedByGenre()` so now the new method also accepts an int N and a float X. This method returns a hashmap with with genres as keys and the highest rated movies as values that have at least N reviews and a rating higher than X. Added some comparators for movies. Added searchers for the movies.